backend/requirements.txt
supabase
fastapi
websockets
python-multipart
pydantic
uvicorn
python-dotenv
pyautogui
opencv-python
ultralytics